refactor: 10.4 예제1 패키지 이동
